+++ src/app.browser.module.ts
@@ -1,24 +1,9 @@
+import { HomeView } from './app.component';
+import { AppComponent } from './app.component';
 import { NgModule, Component } from '@angular/core'
 import { BrowserModule } from '@angular/platform-browser'
 import { RouterModule } from '@angular/router'
 import { APP_BASE_HREF } from '@angular/common';
-
-@Component({
-	selector: 'home-view',
-	template: `<h3>Home View</h3>`
-})
-export class HomeView { }
-
-@Component({
-	selector: 'demo-app',
-	template: `
-	  <h1>Universal Demo</h1>
-	  <a routerLink="/">Home</a>
-	  <a routerLink="/lazy">Lazy</a>
-	  <router-outlet></router-outlet>
-	`
-})
-export class AppComponent { }
 
 @NgModule({
 	imports: [
@@ -36,4 +21,4 @@ export class AppComponent { }
 		{ provide: APP_BASE_HREF, useValue: '/' },
 	]
 })
-export class AppModule { }
+export class AppBrowserModule { }
