+++ src/app.server.module.ts
@@ -1,11 +1,12 @@
 import { NgModule } from '@angular/core';
 import { ServerModule } from '@angular/platform-server';
-import { AppComponent, AppModule } from './app';
+import { AppComponent } from './app.component';
+import { AppBrowserModule } from './app.browser.module';
 
 @NgModule({
   imports: [
 	  ServerModule,
-	  AppModule
+	  AppBrowserModule
   ],
   bootstrap: [
 	  AppComponent
