+++ src/index.html
@@ -5,6 +5,6 @@
 	<base href="/" />
 </head>
 <body>
-  <demo-app></demo-app>
+  <app></app>
 </body>
 </html>
