+++ src/main.browser.ts
@@ -1,8 +1,8 @@
 import './polyfills.ts';
 
 import { platformBrowser } from '@angular/platform-browser'
-import { AppModule } from './app'
-import { AppModuleNgFactory } from './ngfactory/src/app.ngfactory'
+import { AppBrowserModule } from './app.browser.module'
+import { AppBrowserModuleNgFactory } from './ngfactory/src/app.browser.module.ngfactory'
 
 platformBrowser()
-  .bootstrapModuleFactory(AppModuleNgFactory);
+  .bootstrapModuleFactory(AppBrowserModuleNgFactory);
